+++ src/components/timer/index.js
@@ -1,0 +1,4 @@
+import Timer from './Timer';
+
+export default Timer;
+export { Timer };